reorganise code and path aliases
