+++ app/tsconfig.json
@@ -4,10 +4,9 @@
         "strict": true,
         "baseUrl": ".",
         "paths": {
-            "@screens/*": ["./screens/*"],
             "@assets/*": ["./assets/*"],
             "@app/*": ["./app/*"],
-            "@utils/*": ["./utils/*"]
+            "@src/*": ["./src/*"]
         }
     },
     "exclude": ["node_modules", "*.config.js"]
